bll/consumer/requirements.txt
Flask==2.0.2
Flask_SocketIO==5.3.6
python_bcrypt==0.3.2
SQLAlchemy==2.0.30
